Pass all logging through a log function rather than log.Print*().
This is especially useful until we fix the "no takeback" and "no
slide" bugs (which are actually two manifestations of the the same
bug), because in a fact game it's not always possible to notice that
the program has entered an illegal state.

At least with all the position info written to stdout, it's possible
to reconstruct a game after the fact; with the position info written
to stderr it's often lost.