First stab at implementing a one-move lookback for robustness.
The problem is that if the internal game state ever becomes
inconsistent, it never recovers, and game logging just stops.

With these changes, we now try to see if the current position
represents a legal move not only from the most recent accepted move,
but also the one before that.

Testing seems to show that this makes game logging much, much more
robust, especially in fast games. Coindicentally, it also allows for
move takeback (the task of policing 'touch move' should be left to the
human opponent, not the computer).

The first cut implementation is a little ragged and could do with some
refactoring.

Note that this diff also requires the chess.Clone() method, which
isn't in notnil/chess (yet).